Support for new data structures, two-way kpm
Now the new data structures in the issue-tracker are supported. With
these we can sync a single issue to multiple trackers.
Start to support two-way sync with KPM. This now runs through without
repeatedly syncing the same attributes. The sync direction to KPM is not
yet implemented and still has problems.