Refactor to work for jira and roundup backends
Now update_sync_db is fixed to write to the classdict directly. This
ensures that update_aux_classes now has the correct values in the
classdict. Refactor so that update_aux_classes directly calls
update_sync_db and a derived update_aux_classes does a super call.
This also gets rid of an empty implementation of update_aux_classes in
the jira backend.