Sync issues not in mailbox
KPM seems to sometimes not return all issues, e.g. when an issue has
been re-assigned to another supplier. We sync all issues known to us (in
the sync database) that were not found when searching for issues in our
mailbox. This needed refactoring of kpm issue requests into get_problem.