Check readability of issues
Allow use of this readability flag in syncs *and* add a mechanism that
allows to set the remote value to a chosed value if the remote issue is
unreadable. This should allow us to specify, e.g., which status we want
to set the local issue to when the KPM issue is no longer readable for
us.