Fix sync of supplier response
Now if we already synced a supplier response to KPM we later compare the
synced version, not what KPM insists it wanted to mangle the content to.
Also fix the case that the SupplierStatus is '-': If we write this to
KPM we change it to '1', the status '-' should set our local status to.
The reason the sync framework doesn't do this automagically is that the
'-' status is equivalent to '1' and no change is detected.