First stab at remote issue creation with KPM
The local tracker now gets the config as parameter.
Implement filtering of issues for jira.
Implement sync_new_local_issues for jira.
Call sync_new_local_issues in kpmwssync.
Remove id parameter for KPM Problem, was never used and is contained in
the record we pass.
The allowed_actions for a new empty KPM Problem are empty (no remote
actions can be done on a not-yet-existing item)
Add new static (class) attribute to_local to Sync_Attribute classes.
This is True if it doesn't make sense to call the sync when we don't
have a remote issue yet.
Add config option LOCAL_QUERY to filter new local issues to be created
on the remote side. This is a dictionary of (constant) query parameters.
All other filtering should be done in sync config stanzas.